Confirm appropriateness of projections here
In circumstances where population projections and estimates are available
for the same year there is a choice about which is most appropriate. On
balance I favour projections as the demographic multiplier is a percentage
change comparing 2 years from the projection dataset. This seems more
valid than comparing a more recent estimate with an older projection.